chore(docs): make index authoritative for toolingLinks; strip per-file toolingLinks and add cleanup script
